move RowParser to Internal.hs
this addresses https://github.com/Tritlo/duckdb-haskell/issues/6 and
restores functionality to beam-duckdb
